Rename setVanity to vanity, fix vanity util method
