+++ src/shared/data.json
@@ -113,7 +113,7 @@
       "jphelp": [null, "bots", "role-logs"],
       "omega": ["member", "bot", null]
     },
-    "setVanity": {
+    "vanity": {
       "jphelp": "jp101"
     },
     "genLogs": {
